// apps/legacy-api/src/admin/product/enum/product.enum.ts
export enum DateFilterType {
  CURRENT_DATE = 'current_date',
  YESTERDAY = 'yesterday',
  THIS_WEEK = 'this_week',
  LAST_WEEK = 'last_week',
  LAST_7_DAYS = 'last_7_days',
  THIS_MONTH = 'this_month',
  LAST_MONTH = 'last_month',
  THIS_YEAR = 'this_year',
  LAST_YEAR = 'last_year',
  CUSTOM_RANGE = 'custom_range',
  ALL_TIME = 'all_time',
}